RAM banking working

+++ src/globals.c
@@ -24,9 +24,9 @@ BOOLEAN player_alive = TRUE;
 GameState currentGameState = GAME_STATE_NEW_GAME;  // Start with the game loop
 
 // movement
-uint8_t depth = 2, width = 6;
-uint8_t depth_offset = 0, width_offset = 0;
-uint8_t prev_depth, prev_width;
+uint16_t depth = 2, width = 6;
+uint16_t depth_offset = 0, width_offset = 0;
+uint16_t prev_depth, prev_width;
 fixed depth_pixel, width_pixel;
 fixed scroll_x, scroll_y;
